+++ src/Utils.py
@@ -1,0 +1,6 @@
+def isStringInt(string: str) -> bool:
+    try:
+        int(string)
+        return True
+    except ValueError:
+        return False
